Update excludes for new Frechet tests 102 (#103)
* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...


Former-commit-id: f78a6e9c6747d17b0c05494c3969aca508c8ecd1