Merge latest jdk 131 fix into develop (#149)
* Update excludes for new Frechet tests 102 (#103)

* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...

* Fix jdk requires to latest release


Former-commit-id: 678d6fd30004a0cf4f304b26dc57c9a50ae5c5c6