Merge develop into master (#153)
* Merge latest jdk 131 fix into develop (#149)

* Update excludes for new Frechet tests 102 (#103)

* Update test excludes for new Frechet tests

* Fix whitespace

* Update hootenanny.spec.in

Walking through the build process (not fun) led me to this also needing to be changed.  This is where the error is happening.  I think that the other is once the RPMs are installed then it runs the tests again.  This test is creating the RPM for Hootenanny.

* Fixed bad exclude regex

Missing an "S" in the exclude regex...

* Fix jdk requires to latest release

* require npm for services-ui (#137)

* require npm for services-ui

refs #136

* attempt to fix missing failure

* fix bad merge

* Fix for geos updating to 3.5.0 in public repo (#152)


Former-commit-id: af78bde7406fca9fc8d03bb488541f682bfda198